requirements.txt
numpy==1.21.5
Pillow==9.4.0
scikit_image==0.19.2
scikit_learn==1.2.1
scikit_image==0.19.2
streamlit==1.17.0
tensorflow==2.11.0
opencv-python==4.6
altair==4.2.0

